[refactor] 입학년도 선택 컴포넌트 분리
